Fixed up the installer script
Changed the name to setup.py, and removed the batch file (as python will
only ever be on my PATH via powershell in windows)

The new script is a lot cleaner, thanks to Python 3.2. I might
'abstract' it out further and setup the symlink function to take an
iterable and a function that returns the expected format string or what
have you.