After evaluating go, I've decided I won't be using it
While go has some decent ideas, the focus of the language seems to be
for people who love C and python, want higher level ideas, but hate C++.
Unfortunately, I like C++. Add on to it that I just don't know what I
would write it in, with the exception of *maybe* a server side
application, but if I were doing that it'd most likely be for a game
server, and then I'd probably have to start sharing stuff across
boundaries, and doing stuff like serialization and buuuuuuh.

I'll give it a few years, then come back to it. See if anything has
changed.