Add add-registrypath cmdlet to allow registry based env:PATH setting
Fixed the which cmdlet (no idea why it languished the way it did for so
long. Oh well)

Add cmake and 'sliksvn' (my SVN client of choice) to the env:PATH in
this way. Don't think this is an endorsement of SVN. Some people have
not yet seen the light, and sometimes I need to grab their code.