🔥 Fairwell to the old vimrc config o7
My vimrc is now vastly reduced in complexity. It is no longer symlinked
by default, but I'm keeping it around for sentimental reasons, as
strange as that is to say. I've kept this specific file sync'd and
edited and changed across machines for more almost 20 years. I just
can't bring myself to kill it.

📌 Pinned the trouble.nvim plugin to 1.0
♻️ Refactored setup.py to use newer APIs for symlinking

I've started planning on a final way to manage my dotfiles. It'll take
some work and some time, but I think I'll manage.

Signed-off-by: Izzy Muerte <63051+bruxisma@users.noreply.github.com>