🚚 Move to using lazy.nvim over packer.nvim
We're not at 100% parity but it is certainly *good enough* for what I'm
doing. I need to spend some more time configuring additional plugins,
and getting DAP and LSP support to be more refined.

Also need to find a replacement for null_ls now that it's been archived
and is no longer being maintained